requirements.txt
# DB Analyzer AI v5 — Python dependencies

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.0
pydantic-settings==2.0
pyyaml==6.0

# AMAIZ (uncomment with exact package names from your env)
# amaiz_sdk
# amaiz_framework_sdk

# Database connectors
asyncpg==0.29.0
# acouchbase  # Couchbase (optional)

# Multipart form support (required for chat file upload)
python-multipart>=0.0.6

# Auth (required for Keycloak mode)
python-jose[cryptography]==3.3.0
httpx==0.25.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0